Reduce specificity of app.kubernetes.io/part-of label
We will not be able to meaningfully scope and inherit what this Patroni
instance is a part of, other than the common release name. As such, making
the value only have the release name as a pattern makes the most sense as
it will be the common thing binding this deployment along with anything
else that depends on it.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>